Added an id field for all items and delete router
The way the mongoose schema for our shopping list items was originally
set up, there was no id field similar to the field that we used to
have on the in-memory version of the project. There is now an id field
(different from the _id field that is automatically assigned to docs),
and this field will autoincrement through use of the
mongoose-auto-increment module. To get this module working, had to
update the connect.js module. All of the changes routes and services
are for handling DELETE requests from the app.

Closes #2